3주차 상품 조회

+++ week02/src/main/java/com/example/week02/Cart/Cart.java
@@ -25,7 +25,7 @@
 
         @Id
         @GeneratedValue(strategy = GenerationType.IDENTITY)
-        private int cartId;
+        private long cartId;
 
         @Column(nullable = false)
         private LocalDate cartDate;
