+++ week02/src/main/java/com/example/week02/Order/OrderProduct.java
@@ -25,7 +25,7 @@ public class OrderProduct {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int orderProductId;
+    private long orderProductId;
 
     @Column(nullable = false)
     private int quantity;
