+++ week02/src/main/java/com/example/week02/Product/ProductDto.java
@@ -8,7 +8,7 @@ import java.util.stream.Collectors;
 
 // 조회
 public record ProductDto(
-        int productId,
+        long productId,
         String productName,
         String productDesc,
         String image,
